pyproject.toml
[tool.poetry]
name = "bryson-stock"
version = "0.1.0"
description = ""
authors = ["Bryson Xue"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
ta-lib = "^0.4.28"
black = "^24.4.2"
duckdb = "^0.10.2"
sqlalchemy = "^2.0.29"
duckdb-engine = "^0.12.0"
finlab = "^1.1.1"
dbt-duckdb = "^1.7.4"
gspread = "^6.1.2"
oauth2client = "^4.1.3"
openpyxl = "^3.1.3"
python-dotenv = "^1.0.1"
pygsheets = "^2.0.6"
beautifulsoup4 = "^4.12.3"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
